Allow single stocks to parse through buylang
IBKR uses spaces in some stock names like "BRK A" and "BRK B" but our
previous parsing rules couldn't accept those symbol names.

Now we use '_' in place of spaces (but it gets transformed to spaces on
the output logic) so you can "add BRK_A BRK_B" and it works as expected.