Refactor rounder to comply with roundnear updates
Rounding now returns Decimal type and there are some other minor changes
we attempted here, but may not be continued.

We may remove rounder here completely because we replaced it with
instrumentdb over in icil/icli/instrumentdb.py with dynamic live API
price tick lookups for all instruments created on-demand.