Implement orderlang for command line order parsing
The naming is getting confusing here, but the breakdown is:

- orderlang is for FULL ORDER SYNTAX including
  instrument/qty/price/bracket details
- buylang is for a single symbol stock/option/future/spread breakdown

orderlang replaces the manual "buy" command string parsing in icli with
a real parser so we have more flexible conditions now.

Currently, orderlang doesn't use buylang but it probably should. We only
accept single-token instrument names for orderlang to populate, but we
could pass those to buylang to generate an actual underlying order
object in the future.