Fix a bug with orderintent scale directional stops
We were using "regular" math instead of creating the exact sub-types we
use for tagged types when creating the scale order components, so the
final scale order detail didn't retain the original "is long" or "is
short" behavior leading to incorrect estimated stop out prices.

Now we make sure to copy the original class type into the new math for
the scale order components generated.