Add ifthen custom DSL
a much more easier to understand format than manually creating nested
key-value YAML flow control logic.

See the DSL test in the test file for examples of the new predicate name
binding and using predicates and flow control in peer and tree groups.

Basically, peers in the DSL are defined by pipes (a | b | c) since a
peer group is "one cancels all" and trees are defined by arrows since
they are a blocking if-then format (a -> b) but you can also nest and
chain them ((a | b) -> c -> d -> (e | f | g). You can also use '@' as a
final marker to loop the flow control back to the start (a -> b -> @).