Add templating system for `ifthen` auto-generation
The ifthen DSL is great, but trying to manage multiple DSL files across
different combinations (symbols, algos, durations, quantities,
profit/loss targets) gets complicated and messy fast.

So, instead of dsl-in-files, we can now auto-generate DSL content from
templates then consume/auto-start them on-demand.

This also includes a simple "performance tracking and reporting" system
in the new ifthen template system so combinations of ifthen inputs can
be tracked for performance (i.e. if an algo combination is generating
too many losses in a row, we can detect and stop it easily because it
may be counter-firing or just out of sync with current market dynamics).